requirements.txt
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.12.2
django-cors-headers==3.7.0
gunicorn==20.0.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
